CARGO: Run Clippy action
