+++ src/main/kotlin/org/rust/cargo/icons/CargoIcons.kt
@@ -6,4 +6,5 @@ object CargoIcons {
     val ICON = IconLoader.getIcon("/icons/cargo.png")
     val LOCK_ICON = IconLoader.getIcon("/icons/cargo-lock.png")
     val BUILD_RS_ICON = IconLoader.getIcon("/icons/build-rs.png")
+    val CLIPPY = IconLoader.getIcon("/icons/clippy.png")
 }
